refactor: #115 - postpage components index.ts 파일 추가
