+++ src/pages/PostPage/components/index.ts
@@ -1,0 +1,1 @@
+export { default as CommentList } from './CommentList';
